% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidData.R
\docType{data}
\name{covidData}
\alias{covidData}
\title{Our World in Data's Covid Data}
\format{
a tibble with 143825 observations and 5 variables:
\itemize{
\item \strong{xxxx}: full country name
\item \strong{yyyy}: three letter country code
\item \strong{zzzz}: year from 1961 to 2018
\item \strong{wwww}: one of bananas, barley, beans, cassava, cocoa, maize, peas, potatoes, rice, soybeans, wheat
\item \strong{bbbb}: the total yield of the crop, measured in tonnes per hectare
}
}
\source{
\href{https://github.com/owid/covid-19-data/tree/master/public/data}{Our World in Data}
}
\usage{
covidData
}
\description{
A dataset containing crop yields for countries over time. Yields   TODO
are measured in tonnes per hectare for seven crops.                TODO
}
\keyword{datasets}
