% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box.R
\name{text_box}
\alias{text_box}
\title{Text box instructions function}
\usage{
text_box(text)
}
\arguments{
\item{text}{text id}
}
\value{
Box with instructions text set in the server code
}
\description{
Text box instructions function
}
