% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chart.R
\name{data_chart}
\alias{data_chart}
\title{Data processing for charts}
\usage{
data_chart(dataset, group_byparameter1, group_byparameter2)
}
\arguments{
\item{dataset}{This is the dataset to be processed}

\item{group_byparameter1}{This is the 1st parameter of groupby() function}

\item{group_byparameter2}{This is the 2nd parameter of groupby() function}
}
\value{
The dataset processed to be used to plot the chart
}
\description{
Data processing for charts
}
