% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chart.R
\name{data_chart}
\alias{data_chart}
\title{Data processing for charts}
\usage{
data_chart(dataset = covidData, group_byparameter)
}
\arguments{
\item{dataset}{This is the dataset to be processed}

\item{group_byparameter}{This is the column of the dataset to be used inside of groupby() function together with the column "Date"}
}
\value{
The dataset processed to be used to plot the chart
}
\description{
This function prepares the dataset with some filters, groupings and summarizing to be used into plotting codes.
}
\examples{
data_chart(dataset=covidData, group_byparameter="location")

}
