% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box.R
\name{text_box}
\alias{text_box}
\title{Text box instructions function}
\usage{
text_box(text)
}
\arguments{
\item{text}{This is the text id for the text to be written inside the box}
}
\value{
Box with instructions text inside to be set in the server logic unit
}
\description{
This function gives you a box in the wellPanel with text inside
}
\examples{
 
text = "text0"
text_box(text)

}
